test shift

+++ src/testmain/test2.go
@@ -52,7 +52,7 @@ func NewCat(name string) Cat {
 
 func test() {
 	num := [...]uint8{1, 2, 3, 4, 5, 6}
-	switch 2 {
+	switch uint8(2) {
 	case num[1], num[2]:
 		fmt.Println("xx")
 		// fallthrough
